feat(mendix-widgets-mcp): dockerize the server
