+++ mcp/.gitignore
