+++ mcp/src/diff-engine.ts
