+++ mcp/src/prompts.ts
