+++ mcp/src/property-engine.ts
