+++ mcp/tsconfig.json
